+++ tests/gateway/test_discord_component_auth.py
@@ -18,7 +18,7 @@ import pytest
 
 # Trigger the shared discord mock from tests/gateway/conftest.py before
 # importing the production module.
-from gateway.platforms.discord import (  # noqa: E402
+from plugins.platforms.discord.adapter import (  # noqa: E402
     ExecApprovalView,
     ModelPickerView,
     SlashConfirmView,
