+++ tests/gateway/test_discord_media_metadata.py
@@ -1,6 +1,6 @@
 import inspect
 
-from gateway.platforms.discord import DiscordAdapter
+from plugins.platforms.discord.adapter import DiscordAdapter
 
 
 def test_discord_media_methods_accept_metadata_kwarg():
