+++ tests/gateway/test_discord_model_picker.py
@@ -11,7 +11,7 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from gateway.platforms.discord import ModelPickerView
+from plugins.platforms.discord.adapter import ModelPickerView
 
 
 @pytest.mark.asyncio
